+++ src/services/linkService.js
@@ -1,5 +1,5 @@
 import Database from './gateway/database'
-import { Messaging } from './gateway/messaging'
+import { PushNotificationService } from './pushService'
 
 export default function LinkService() {
   const checkLink = (linkingUser, linkedUser) => {
@@ -16,21 +16,7 @@ export default function LinkService() {
           matchesToCreate[`${linkingUser}/${linkedUser}/read`] = false
           return matchesRef.update(matchesToCreate).then(() => {
             console.log('\tMatch successfully created!')
-            const payload = {
-              notification: {
-                title: 'Nuevo match!',
-                body: 'Esta es una push notification creada desde el servidor',
-                icon: 'myicon'
-              }
-            }
-            const registrationToken = 'cNgfu6nRls0:APA91bGaNy7PJ8xfYPQCUvvEOtSelBzTb6cpx9JvX0woGsm19-pwUC5jrMtCsIwjYOM-sMhDaRI_bNHVUCEP2Svyp0u3eLJ9POz-PZJymh7JKHIgZjYdilo2SB1XfiFAbxKCXjg92Uic'
-            Messaging().sendToDevice(registrationToken, payload)
-              .then(response => {
-                console.log('Successfully sent message:', response)
-              })
-              .catch(error => {
-                console.log('Error sending message:', error)
-              })
+            PushNotificationService().sendMatchPush()
           }).catch(() => {
             console.log('\tMatch could not be created :(')
           })
